+++ ui/src/lib/utils/theme.ts
@@ -2,11 +2,10 @@ import { isMobile, isTauriEnv } from '$lib/utils/environment';
 
 type BarStyle = 'light' | 'dark';
 
-export async function setSystemBarsStyle(
+async function setSystemBarsStyle(
 	statusBarStyle: BarStyle,
 	navigationBarStyle: BarStyle,
 ) {
-	if (!isMobile || !isTauriEnv()) return;
 	const { invoke } = await import('@tauri-apps/api/core');
 	await invoke('plugin:system-bars-styles|set_style', {
 		statusBarStyle,
@@ -15,8 +14,19 @@ export async function setSystemBarsStyle(
 	});
 }
 
+/**
+ * Force light (white) system bars, for the always-dark overlays (lightbox,
+ * staged-media page) whose background is dark regardless of the app theme.
+ */
+export async function setLightSystemBars() {
+	if (isMobile && isTauriEnv()) {
+		await setSystemBarsStyle('light', 'light');
+	}
+}
+
 /** Restore the system-bar style to match the currently-applied theme. */
 export async function applyThemeSystemBars() {
+	if (!isMobile || !isTauriEnv()) return;
 	const dark = document.documentElement.classList.contains('dark');
 	await setSystemBarsStyle(dark ? 'light' : 'dark', dark ? 'light' : 'dark');
 }
